i dont remember what i did

+++ shop/models.py
@@ -7,7 +7,7 @@ class Coffee(models.Model):
     id = models.UUIDField(primary_key=True,default=uuid.uuid4,editable=False)
     name = models.CharField(max_length=255)
     description = models.TextField(blank=True,null=True)
-    price = models.PositiveIntegerField()
+    price = models.DecimalField(max_digits=10,decimal_places=2)
     available = models.BooleanField(default=True)
     stock = models.PositiveIntegerField()
     cover = models.ImageField(upload_to='coffee_covers/',blank=True)
